+++ src/org/ods/services/NexusService.groovy
@@ -49,7 +49,8 @@ class NexusService {
                 passwordVariable: 'PASSWORD'
             )
         ]) {
-            new NexusService(c.nexusUrl as String, script.env.USERNAME as String, script.env.PASSWORD as String)
+            new NexusService(c.nexusUrl as String, script.env.USERNAME as String, 
+                script.env.PASSWORD as String)
         }
     }