+++ vars/odsComponentStageScanWithAqua.groovy
@@ -54,7 +54,8 @@ def call(IContext context, Map config = [:]) {
                 passwordVariable: 'PASSWORD'
             )
         ]) {
-            nexusService = new NexusService(context.nexusUrl as String, this.env.USERNAME as String, this.env.PASSWORD as String)
+            nexusService = new NexusService(context.nexusUrl as String, this.env.USERNAME as String,
+                this.env.PASSWORD as String)
         }
         registry.add(NexusService, nexusService)
     }