+++ vars/odsComponentStageScanWithSonar.groovy
@@ -40,7 +40,8 @@ def call(IContext context, Map config = [:]) {
                 passwordVariable: 'PASSWORD'
             )
         ]) {
-            nexusService = new NexusService(context.nexusUrl, this.env.USERNAME as String, this.env.PASSWORD as String)
+            nexusService = new NexusService(context.nexusUrl, this.env.USERNAME as String, 
+                this.env.PASSWORD as String)
         }
     }
     def stage = new ScanWithSonarStage(